+++ Prefabs/Code/MountainKitWoodGate.cs
@@ -1,6 +1,7 @@
 ﻿using A2.CleanLargeWoodGate.Unity;
 using System.Collections.Generic;
 using UnityEngine;
+using static A2.CleanLargeWoodGate.Prefabs.PrefabTools;
 
 namespace A2.CleanLargeWoodGate.Prefabs.Code
 {
@@ -9,67 +10,32 @@ namespace A2.CleanLargeWoodGate.Prefabs.Code
         public const string Name = "Mountain kit wood gate";
         public const string PrefabName = "MountainKit_wood_gate";
 
-        public static bool Modify(Dictionary<string, GameObject> prefabs)
+        public static bool Modify(IReadOnlyDictionary<string, GameObject> prefabs, IReadOnlyDictionary<string, GameObject[]> clones)
+            => TryModify(prefabs, clones, PrefabName, ref Flags.MountainKitWoodGate, Modify, nameof(MountainKitWoodGate), nameof(Modify));
+        public static bool Restore(IReadOnlyDictionary<string, GameObject> prefabs, IReadOnlyDictionary<string, GameObject[]> clones)
+            => TryRestore(prefabs, clones, PrefabName, ref Flags.MountainKitWoodGate, Restore, nameof(MountainKitWoodGate), nameof(Restore));
+
+        private static bool Modify(GameObject prefab)
         {
-            try
-            {
-                if (Flags.MountainKitWoodGate != PrefabState.ToModify) return false;
-#if DEBUG
-                Jotunn.Logger.LogInfo($"{nameof(MountainKitWoodGate)}.{nameof(Modify)}: modifying state of the prefab {PrefabName}");
-#endif
-                if (!prefabs.TryGetValue(PrefabName, out var prefab))
-                {
-                    Jotunn.Logger.LogInfo($"{nameof(MountainKitWoodGate)}.{nameof(Modify)}: Prefab {PrefabName} not found.");
-                    return false;
-                }
-
-                var result = true;
-                result = prefab.SetChildrenInactive(
-                    "CastleKit_decal_fenrir_blood",
-                    "CastleKit_decal_fenrir_blood (1)",
-                    "CastleKit_decal_fenrir_blood (2)",
-                    "CastleKit_decal_fenrir_blood (3)"
-                    ) && result;
-
-                if (result) Flags.MountainKitWoodGate = PrefabState.Modified;
-                return result;
-            }
-            catch (System.Exception ex)
-            {
-                Jotunn.Logger.LogError($"{nameof(MountainKitWoodGate)}.{nameof(Restore)}: Exception occurred:\n{ex}");
-                return false;
-            }
+            var result = true;
+            result = prefab.SetChildrenInactive(
+                "CastleKit_decal_fenrir_blood",
+                "CastleKit_decal_fenrir_blood (1)",
+                "CastleKit_decal_fenrir_blood (2)",
+                "CastleKit_decal_fenrir_blood (3)"
+                ) && result;
+            return result;
         }
-        public static bool Restore(Dictionary<string, GameObject> prefabs)
+        private static bool Restore(GameObject prefab)
         {
-            try
-            {
-                if (Flags.MountainKitWoodGate != PrefabState.ToRestore) return false;
-#if DEBUG
-                Jotunn.Logger.LogInfo($"{nameof(MountainKitWoodGate)}.{nameof(Restore)}: restoring state of the prefab {PrefabName}");
-#endif
-                if (!prefabs.TryGetValue(PrefabName, out var prefab))
-                {
-                    Jotunn.Logger.LogInfo($"{nameof(MountainKitWoodGate)}.{nameof(Restore)}: Prefab {PrefabName} not found.");
-                    return false;
-                }
-
-                var result = true;
-                result = prefab.SetChildrenActive(
-                    "CastleKit_decal_fenrir_blood",
-                    "CastleKit_decal_fenrir_blood (1)",
-                    "CastleKit_decal_fenrir_blood (2)",
-                    "CastleKit_decal_fenrir_blood (3)"
-                    ) && result;
-
-                if (result) Flags.MountainKitWoodGate = PrefabState.Restored;
-                return result;
-            }
-            catch (System.Exception ex)
-            {
-                Jotunn.Logger.LogError($"{nameof(MountainKitWoodGate)}.{nameof(Restore)}: Exception occurred:\n{ex}");
-                return false;
-            }
+            var result = true;
+            result = prefab.SetChildrenActive(
+                "CastleKit_decal_fenrir_blood",
+                "CastleKit_decal_fenrir_blood (1)",
+                "CastleKit_decal_fenrir_blood (2)",
+                "CastleKit_decal_fenrir_blood (3)"
+                ) && result;
+            return result;
         }
     }
 }